Phew...stage 6! Big one!
Ok so..i almost succeeded in merging this with auto scroll animations and their intergration
with manual control!

Problem was html skeleton. In previous iteration it made it impossible at final stages -
the switch of control from animation to manual controls. This html skeleton overhaul and
following css adjustments serve to make the full animation implementation a possibility