Animations are here!
So, remember that big ass failed attempt? Well 90% of it was good.
One infinite repeat setting for animations, one removal of marking active nav button styles,
do it does not collide with animations and voila!

It works and synergizes with one another (manual operation with automatic animations)
Only thing left are animations for arrow navigation as these are still outta sync.